SERVICE : Add Avarage Stars Buyer Post

+++ src/GreenSale.DataAccess/ViewModels/BuyerPosts/BuyerPostViewModel.cs
@@ -22,7 +22,7 @@ namespace GreenSale.DataAccess.ViewModels.BuyerPosts
         public string District { get; set; } = string.Empty;
         public string Address { get; set; } = string.Empty;
         public BuyerPostEnum Status { get; set; }
-        public long AverageStars { get; set; }
+        public double AverageStars { get; set; }
         public int UserStars { get; set; } 
         public DateTime CreatedAt { get; set; }
         public DateTime UpdatedAt { get; set; }
